+++ packages/terra-core-docs/src/terra-dev-site/test/heading/DefaultHeading.test.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import Heading from 'terra-heading';
+
+export default () => <Heading id="heading-default" level={1}>Default</Heading>;