+++ packages/terra-heading/src/terra-dev-site/test/heading/DefaultHeading.test.jsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import Heading from '../../../Heading';
-
-export default () => <Heading id="heading-default" level={1}>Default</Heading>;