+++ packages/terra-hyperlink/src/terra-dev-site/test/hyperlink/DefaultHyperlink.test.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import Hyperlink from '../../../Hyperlink';
-
-export default () => (
-  <div role="main">
-    <Hyperlink id="link" href="https://www.cerner.com">Default hyperlink</Hyperlink>
-  </div>
-);